app/requirements.txt
torch
torchvision
pytorchcv
facenet-pytorch
efficientnet-pytorch
opencv-python-headless
pandas
numpy
Pillow
